moved some constants to a common file
