+++ api/app/geospatial.py
@@ -1,0 +1,12 @@
+""" Geospatial things
+"""
+from typing import Final
+from pyproj import CRS
+
+# Some constants that are frequently used when transforming coordinates.
+
+# BCGOV standard is to store everything in NAD83
+NAD83: Final = 'epsg:4269'
+NAD83_CRS: Final = CRS(NAD83)
+# De facto standard is to expose data in WGS84
+WGS84: Final = 'epsg:4326'
